fix: move signer to signer directory
